+++ crates/bif_qt/cpp/property_inspector_widget.h
@@ -18,6 +18,7 @@
 #include <QWidget>
 
 class BifShellState;
+class QCheckBox;
 class QGroupBox;
 class QLabel;
 class QStandardItemModel;
@@ -35,6 +36,7 @@ public:
 private slots:
     void on_selection_changed();
     void on_layer_state_changed();
+    void on_visibility_toggled(bool checked);
 
 private:
     void rebuild();
@@ -45,6 +47,7 @@ private:
 
     QLabel* m_header_path;
     QLabel* m_header_type;
+    QCheckBox* m_visibility_box;
     QGroupBox* m_arcs_group;
     QListWidget* m_arcs_list;
     QTabWidget* m_tabs;